python/pyproject.toml
[project]
name = "aguitest"
version = "0.1.0"
description = "Learning AGUI (Agent User Interaction Protocol) SDK"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ag-ui-protocol>=0.1.10",
    "fastapi>=0.115.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "pydantic-ai-slim[ag-ui,bedrock,google,openai]>=1.31.1",
    "pillow>=11.0.0",
    "simpleeval>=1.0.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.389",
    "ruff>=0.14.1",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
    "pytest-cov>=6.0.0",
    "pytest-playwright>=0.6.0"
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
# Ignore DeprecationWarning from google-genai 1.68.0 upstream bug
filterwarnings = [
    "ignore:'_UnionGenericAlias' is deprecated:DeprecationWarning",
]

[tool.ruff.lint]
extend-select = ["PLC0415"]
